moved RF coputation back into tuning.py
